BLZ1 Rework + Hiatus Build
Bridge Lagoon 1 is Ready for Playtesting!

Oh and I replaced BLZ1 and SGZ1's PH Music and made more of the songs loop so that's nice.

Oh and i'm going on a Short Hiatus on Sock, maybe beside BGs i'm not going to touch the project for a bit of time so i can decompress and not Cancel the Project on the spot or anything.
Thank you for your understanding and have a good night/day!